Backported most of the TTC implementation of EMF queries

+++ hu.bme.mit.trainbenchmark.generator.graph/src/main/java/hu/bme/mit/trainbenchmark/generator/graph/GraphGenerator.java
@@ -75,7 +75,7 @@ public class GraphGenerator extends Generator {
 	}
 
 	@Override
-	protected Object createVertex(final long id, final String type, final Map<String, Object> attributes,
+	protected Object createVertex(final int id, final String type, final Map<String, Object> attributes,
 			final Map<String, Object> outgoingEdges, final Map<String, Object> incomingEdges) {
 		final Node node = graphDb.createNode(DynamicLabel.label(type));
