+++ hu.bme.mit.trainbenchmark.generator.sql/src/main/java/hu/bme/mit/trainbenchmark/generator/sql/SQLGenerator.java
@@ -79,7 +79,7 @@ public class SQLGenerator extends Generator {
 	}
 
 	@Override
-	protected Object createVertex(final long id, final String type, final Map<String, Object> attributes,
+	protected Object createVertex(final int id, final String type, final Map<String, Object> attributes,
 			final Map<String, Object> outgoingEdges, final Map<String, Object> incomingEdges) throws IOException {
 		final StringBuilder columns = new StringBuilder();
 		final StringBuilder values = new StringBuilder();
